+++ tests/DeploymentTest.php
@@ -3,7 +3,6 @@
 namespace Laravel\Tests\Forge;
 
 use Closure;
-use InvalidArgumentException;
 use Laravel\Forge\Sites\Site;
 use PHPUnit\Framework\TestCase;
 use Laravel\Tests\Forge\Helpers\Api;